requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
qrcode[pil]>=7.0
pillow>=10.0.0